pyproject.toml
[tool.poetry]
name = "reddit"
version = "0.1.0"
description = ""
authors = ["alanshkim <alanshkim@gmail.com>"]

[tool.poetry.dependencies]
python = "=3.9.1"
pandas = "^2.2.3"
praw = "=7.7.1"
configparser = "^7.1.0"
pyyaml = "^6.0.2"
cryptography = "^43.0.3"
apache-airflow = "^2.10.2"
s3fs = "^2024.10.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
